Move defineClauses to separate module
